requirements.txt
flask==3.0.3
flask-socketio==5.3.6
flask-cors==4.0.1
python-socketio==5.11.2
gunicorn==22.0.0
gevent==24.2.1
gevent-websocket==0.10.1
psycopg2-binary==2.9.11
